+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.4.0
+
+* added custom event tracking
+* added order by unique visitors to conversion goals
+* fixed relative visitors/views and conversion rate if right side is zero
+* updated User-Agent blacklist
+* updated dependencies
+
 ## 2.3.0
 
 * added `MaxTimeOnPageSeconds` option to filter